doc(sanity): add TS-doc for block content related types (#4674)
* chore(core/form/inputs): remove outdated workshop case

Remove this non-working workshop left over from v2

* refactor(core/form/inputs): deprecate block marker callbacks and support

* doc(types): add ts doc comments to block schema types

* doc(sanity): add TS-doc for block content related types
