+++ packages/sanity/src/core/components/previews/types.ts
@@ -24,7 +24,7 @@ export type PreviewLayoutKey = GeneralPreviewLayoutKey | PortableTextPreviewLayo
 
 /**
  * @hidden
- * @beta
+ * @public
  */
 export interface PreviewMediaDimensions {
   aspect?: number
