+++ packages/sanity/src/core/form/inputs/PortableText/__workshop__/customSchema/blockActions.tsx
@@ -3,7 +3,7 @@ import React, {memo, useCallback} from 'react'
 import {CopyIcon} from '@sanity/icons'
 import {keyGenerator} from '@sanity/portable-text-editor'
 import {PortableTextBlock, PortableTextTextBlock} from '@sanity/types'
-import {RenderBlockActionsCallback} from '../../types'
+import {RenderBlockActionsCallback} from '../../../../types/_transitional'
 
 const BlockActions = memo(function BlockActions(props: {
   block: PortableTextBlock
