--- packages/sanity/src/core/form/inputs/PortableText/__workshop__/editObjects/index.ts
@@ -1,1 +1,0 @@
-export {EditObjectsStory as default} from './EditObjectsStory'
