+++ packages/sanity/src/core/form/types/fieldProps.ts
@@ -28,7 +28,7 @@ import {
 
 /**
  * @hidden
- * @beta */
+ * @public */
 export interface BaseFieldProps {
   /** @beta */
   actions?: DocumentFieldAction[]
@@ -55,7 +55,7 @@ export interface BaseFieldProps {
 
 /**
  * @hidden
- * @beta */
+ * @public */
 export interface ObjectFieldProps<T = Record<string, any>> extends BaseFieldProps {
   schemaType: ObjectSchemaType
   value: {[field in string]: unknown} | undefined
@@ -71,7 +71,7 @@ export interface ObjectFieldProps<T = Record<string, any>> extends BaseFieldProp
 
 /**
  * @hidden
- * @beta */
+ * @public */
 export interface ArrayFieldProps extends BaseFieldProps {
   schemaType: ArraySchemaType
   value: unknown[] | undefined
@@ -84,7 +84,7 @@ export interface ArrayFieldProps extends BaseFieldProps {
 
 /**
  * @hidden
- * @beta */
+ * @public */
 export interface ArrayOfPrimitivesFieldProps extends BaseFieldProps {
   schemaType: ArraySchemaType
   value: unknown[] | undefined
@@ -97,7 +97,7 @@ export interface ArrayOfPrimitivesFieldProps extends BaseFieldProps {
 
 /**
  * @hidden
- * @beta */
+ * @public */
 export interface NumberFieldProps extends BaseFieldProps {
   schemaType: NumberSchemaType
   value: number | undefined
@@ -106,7 +106,7 @@ export interface NumberFieldProps extends BaseFieldProps {
 
 /**
  * @hidden
- * @beta */
+ * @public */
 export interface BooleanFieldProps extends BaseFieldProps {
   schemaType: BooleanSchemaType
   value: boolean | undefined
@@ -115,7 +115,7 @@ export interface BooleanFieldProps extends BaseFieldProps {
 
 /**
  * @hidden
- * @beta */
+ * @public */
 export interface StringFieldProps extends BaseFieldProps {
   schemaType: StringSchemaType
   value: string | undefined
@@ -127,7 +127,7 @@ export type PrimitiveFieldProps = NumberFieldProps | BooleanFieldProps | StringF
 
 /**
  * @hidden
- * @beta */
+ * @public */
 export type FieldProps =
   | ObjectFieldProps
   | ObjectFieldProps<CrossDatasetReferenceValue>
